submodule renaming
